# openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Spring Shop Flyway API
  description: |
    # Online Bookstore API
    아마존과 유사한 온라인 도서 주문 및 배송 서비스 API입니다.

    <details>
    <summary><strong>Error Codes (Click to expand)</strong></summary>

    | Code | Message | Status | Description |
    |---|---|---|---|
    | `C001` | Invalid Input Value | 400 | 입력값이 올바르지 않습니다. |
    | `C002` | Entity Not Found | 404 | 요청한 리소스를 찾을 수 없습니다. |
    | `M001` | Email Already Exists | 409 | 이미 가입된 이메일입니다. |
    | `M002` | Login Failed | 401 | 이메일 또는 비밀번호가 일치하지 않습니다. |
    | `B001` | Out of Stock | 409 | 재고가 부족합니다. |
    | `O001` | Order Not Found | 404 | 주문을 찾을 수 없습니다. |
    | `D001` | Delivery Not Found | 404 | 배송 정보를 찾을 수 없습니다. |
    | `R001` | Review Not Allowed | 403 | 구매 이력이 없는 도서에는 리뷰를 작성할 수 없습니다. |
    | `R002` | Review Already Exists | 409 | 이미 리뷰를 작성한 도서입니다. |

    </details>

  version: 0.0.1
servers:
  - url: /api/v1
    description: Relative Path (Environment Independent)
...
paths:
...
  /cart:
    get:
      tags:
        - Cart
      summary: 장바구니 조회
      operationId: getCart
      security:
        - bearerAuth: []
      responses:
        '200':
          description: 조회 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CartResponse'
  /cart/items:
    post:
      tags:
        - Cart
      summary: 장바구니 담기
      operationId: addCartItem
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddCartItemRequest'
      responses:
        '201':
          description: 추가 성공
        '400':
          description: 잘못된 입력
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 도서 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /cart/items/{bookId}:
    put:
      tags:
        - Cart
      summary: 장바구니 수량 수정
      operationId: updateCartItem
      security:
        - bearerAuth: []
      parameters:
        - name: bookId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCartItemRequest'
      responses:
        '200':
          description: 수정 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CartResponse'
        '400':
          description: 잘못된 입력
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 장바구니 항목 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags:
        - Cart
      summary: 장바구니 항목 삭제
      operationId: deleteCartItem
      security:
        - bearerAuth: []
      parameters:
        - name: bookId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: 삭제 성공
...
  /wishlist:
    get:
      tags:
        - Wishlist
      summary: 위시리스트 목록 조회
      operationId: getWishlist
      security:
        - bearerAuth: []
      responses:
        '200':
          description: 조회 성공
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/WishlistItemResponse'
  /wishlist/books/{bookId}:
    put:
      tags:
        - Wishlist
      summary: 위시리스트 추가
      operationId: addWishlistBook
      security:
        - bearerAuth: []
      parameters:
        - name: bookId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: 추가 성공
        '404':
          description: 도서 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags:
        - Wishlist
      summary: 위시리스트 삭제
      operationId: deleteWishlistBook
      security:
        - bearerAuth: []
      parameters:
        - name: bookId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: 삭제 성공
...
components:
  securitySchemes:
    bearerAuth:
      type: oauth2
      description: Keycloak OAuth2 Authorization Code Flow (Google SSO via Keycloak)
      flows:
        authorizationCode:
          authorizationUrl: https://accounts-dev.intelliantech.com/realms/intellian-app/protocol/openid-connect/auth
          tokenUrl: https://accounts-dev.intelliantech.com/realms/intellian-app/protocol/openid-connect/token
          scopes:
            openid: OpenID Connect scope
            profile: Access user profile
            email: Access user email
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        code:
          type: string
          example: "C001"
        message:
          type: string
          example: "Invalid Input Value"
        status:
          type: integer
          example: 400

    # --- Member ---
...
    AddCartItemRequest:
      type: object
      required:
        - bookId
        - quantity
      properties:
        bookId:
          type: integer
          format: int64
        quantity:
          type: integer
          minimum: 1
    UpdateCartItemRequest:
      type: object
      required:
        - quantity
      properties:
        quantity:
          type: integer
          minimum: 1
    CartItemResponse:
      type: object
      properties:
        bookId:
          type: integer
          format: int64
        title:
          type: string
        author:
          type: string
        unitPrice:
          type: integer
        quantity:
          type: integer
        subtotal:
          type: integer
    CartResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/CartItemResponse'
        totalQuantity:
          type: integer
        totalAmount:
          type: integer

    # --- Review ---
...
    WishlistItemResponse:
      type: object
      properties:
        bookId:
          type: integer
          format: int64
        title:
          type: string
        author:
          type: string
        price:
          type: integer

    # --- Order ---
...